Solving a warning

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -16,7 +16,7 @@ return (0);
 
 va_start(args, n);
 
-for (i = 0; i < n; i++)
+for (i = 0; i < (int)n; i++)
 sum += va_arg(args, int);
 
 va_end(args);
